requirements.txt
py-ballisticcalc==2.1.0
py-ballisticcalc[exts]
streamlit
plotly
python-dotenv
openai
